make MySQL auto close statement for batching and minor docs
